+++ CHANGELOG.md
@@ -1,6 +1,4 @@
 # Cynosure `@VERSION@` `1.20.1` `Fabric & Forge`
 
 ### Changelog:
-- 0.1.12 never existed in the first place
-- Fixed crash with satin
-- POI modification now also works on forge
+- New Entity Render Layer registration allowing for abstract renderers